Move changes back to main.ts
Makes reviewing this PR easier by undoing the refactor.
